Allow EEPROM versioning to upgrade to latest (#347)
This change introduces a mechanism to preserve older versions of the
settings in code, with a process to upgrade to the current version on
flash.

It also allows for arbitrary manipulation of the variables during
version bumps, for instance, to transform a stored data type or data
units, or to selectively override some stored settings with defaults. I
expect this to help changes such as PZ range.

It makes the assumption that versions always increase. It does not
handle versions 1-3, as I tried to reconstruct them from git history,
but couldn't find a consistent schema for most of them.